requirements.txt
fastapi
uvicorn
google-cloud-storage
diffusers
torch
pillow
transformers
accelerate 
scipy 
safetensors